TropicalHomotopies.jl

Tropical homotopies from intersection data

TropicalHomotopies.homotopy_from_tropical_data — Method

homotopy_from_tropical_data(G,w)

Given a polynomial system G and a rational point w in the format of OSCAR, this function computes a homotopy in the format of OSCAR using the main algorithm in the paper.

Note

This function is not yet harmonized with polyhedral_homotopies_and_start_systems.

Tropical solving

TropicalHomotopies.HC_system_from_Oscar_system — Method

HC_system_from_Oscar_system(polynomial_system)

Convert OSCAR system (in the form of a system over Q) to a System object in the HomotopyContinuation.jl package.

The variables in the output will always be named x.

TropicalHomotopies.export_homotopy_from_oscar_to_HC — Method

export_homotopy_from_oscar_to_HC(homotopy)

Convert a homotopy from OSCAR (in the form of a system over the rational function field Q(t)) to aHomotopy object in the HomotopyContinuation.jl package.

The variables in the output will always be named x, and the time variable will always be named t.

TropicalHomotopies.parametric_HC_system_from_parametric_Oscar_system — Method

parametric_HC_system_from_parametric_Oscar_system(polynomial_system)

Convert a parametric system (in the form of a system over Q[a]) to a parametrized System object in the HomotopyContinuation.jl package.

The parameters in the output will always be named a and the variables will always be named x.

TropicalHomotopies.solve_binomial_system — Method

solve_binomial_system(F::HC.ModelKit.System)

Solve a binomial system of equations (i.e., a system where each polynomial has two terms) using the BinomialSystemSolver from the HomotopyContinuation.jl package.

TropicalHomotopies.tropical_solve — Method

tropical_solve(F::Vector{<:PolyRingElem},target_parameters::Vector{QQFieldElem},type_of_system::Symbol;kwargs...)

Solve a system of polynomial equations with our tropical techniques. The input consists of a parametrized system F, a vector of target parameters target_parameters, and the type of system type_of_system.

The target_parameters should be generic and rational. (One can use a parameter homotopy to trace the solutions to a specific choice of parameters later.)

Note

As of now, type_of_system has to be :vertical. Implementing :horizontal is work in progress.

Vertical systems

TropicalHomotopies.modify_vertically — Method

modify_vertically(F::Vector{<:MPolyRingElem})

Modify a vertically parametrized system into a linear part and a binomial part.

All the parameters will be in the binomial part.

The modification variables will be denoted by y[1],...,y[m], where m is the number of distinct monomials in the input system.

TropicalHomotopies.tropical_root_count_with_homotopy_data_vertical — Method

tropical_root_count_with_homotopy_data_vertical(F; perturbed_parameters=nothing)

Compute tropical intersection data for a vertically parametrized system through a vertical modification.

If no choice of parameters perturbed_parameters is given, random ones will be chosen.

The output will be a tuple consisting of:

The generic root count

The points in the tropical stable intersection

The initial system for each point

The tropical Gröbner bases for each point

The choice of perturbed_parameters used for the computation.

Warning

If the choice of perturbed_parameters is not generic enough, an error will be thrown. In this case, you should try again with a different choice of perturbed_parameters.

TropicalHomotopies.vertical_embedding — Method

vertical_embedding(F::Vector{<:MPolyRingElem})

Embedd a polynomial system over QQ into a vertically parametrized family in the canonical way.

The variables in the output will be the same as in the input system.

The parameters will be a[1],...,a[m], where m is the number of distinct monomials in the input system.

Polyhedral homotopies

TropicalHomotopies.polyhedral_homotopies_and_start_systems — Method

polyhedral_homotopies_and_start_systems(F::HomotopyContinuation.ModelKit.System)

Compute polyhedral homotopies and start systems for a system of polynomial equations in the format of the HomotopyContinuation.jl package. This is done by computing a mixed subdivision for a randomly chosen lifting. We obtain one polyhedral homotopy and one start system for each mixed cell.

Helper functions

TropicalHomotopies.tropical_stable_intersection_linear_binomial — Method

" tropicalstableintersectionlinearbinomial(TropL::TropicalLinearSpace,TropB::TropicalVariety)

Specialized stable intersection function for a tropical linear space and a linear space (encoded as a tropicalization of a binomial variety).

The output is a vector of stable intersection points and a vector with the multiplicities of the points.

TropicalHomotopies.coefficient_matrix_and_monomials — Method

coefficient_matrix_and_monomials(F::Vector{<:MPolyRingElem})

Extract a coefficient matrix and a vector of monomials for a system F.

TropicalHomotopies.specialize — Method

specialize(F::Vector{<:MPolyRingElem}, choice_of_parameters::Vector{<:Union{Int, RingElem}})

Specialize a parametric polynomial system F at a choice_of_parameters